+++ src/components/pages/Blog/Blog.tsx
@@ -1,11 +1,8 @@
-import { Link } from '#components/common/Links'
 import { Subscribe } from '#components/common/Subscribe'
 import { Metadata, FCC } from '#types/types'
-import { MDXRemote } from 'next-mdx-remote'
 import { useState } from 'react'
-import { RandomColors } from 'src/helpers/colors'
-import { TimestampToDate } from 'src/helpers/TimestampToDate'
 import { useColorContext } from '../../templates/Providers/ColorProvider'
+import { BlogItem } from './BlogItem'
 import { Filter } from './Filter'
 
 export type PostProps = Array<{
@@ -16,23 +13,6 @@ export type PostProps = Array<{
 export type BlogProps = {
   posts: PostProps
   slug?: string
-}
-
-const generateTitle = (compiledTitle: string, colors: RandomColors) => {
-  const text = colors.textColor ?? ''
-
-  const titleComponents = {
-    p: (props: JSX.IntrinsicElements['p']) => (
-      <h2 {...props}>{props.children}</h2>
-    ),
-    strong: (props: JSX.IntrinsicElements['strong']) => (
-      <strong {...props} className={`${text}`} />
-    ),
-  }
-
-  return (
-    <MDXRemote compiledSource={compiledTitle} components={titleComponents} />
-  )
 }
 
 export const Blog: FCC<BlogProps> = ({ posts }) => {
@@ -80,28 +60,7 @@ export const Blog: FCC<BlogProps> = ({ posts }) => {
       <Filter order={order} handleOrder={handleOrder} />
       <ul className="flex flex-col gap-16">
         {blogPosts.map(({ slug, metadata }) => {
-          return (
-            <li key={slug} className="relative pl-6">
-              <div
-                className={`absolute top-0 left-0 w-2 h-full ${colors.bgColor}`}
-              />
-              {metadata?.compiledTitle && (
-                <Link inner href={`/blog/${slug}`}>
-                  {generateTitle(metadata.compiledTitle, colors)}
-                </Link>
-              )}
-              <p className="mb-0">
-                {metadata.description ?? ''}
-                {metadata.createdAt && (
-                  <span className="ml-2">
-                    <small className="opacity-80">
-                      {TimestampToDate(metadata.createdAt)}
-                    </small>
-                  </span>
-                )}
-              </p>
-            </li>
-          )
+          return <BlogItem key={slug} metadata={metadata} slug={slug} />
         })}
       </ul>
     </>
